requirements.txt
requests
supabase
python-dotenv
pyyaml
resend
